Add a plugin to support native ladder tour tracking
Porting this ov over from my bot. It's used by an assload of rooms and basically every major tier, but sucks to maintain as a bot. Will be a much nicer experience as a plugin, and people have asked frequently if it could just be integrated into PS, so... here we are.