Daily Spotlight: Delete button in daily command page
https://www.smogon.com/forums/threads/delete-daily-button-in-the-page.3742300/#post-10533943 
Suggestion is self explanatory but tricky, mostly because of que daily's but I got it to work after some hours but there still remains a small quirk that's not actually my fault but moreso an issue of the removedaily command. You can't actually delete an entire daily if there are que'd daily's to it, I guess this makes sense to prevent accidents if you have tons of dailies que'd but what this means is you can never delete dailies mass at once. So for example if you were to have a set daily with 10 que's on it, you cannot delete the entire daily at once with the delete button which also means the /removedaily command. Currently with the button in the menu, it'll delete singular dailies and que'd dailies specific to the que you want to delete. Yet if there are multiple que's for a daily and you try to delete the original non que daily it'll just delete the most recently added queue daily. 

HOWEVER this is how the command ACTUALLY works when you do /removedaily (not the removeque), so I mean the button works as 'intended' but the command to removedaily's is a bit weird.