Bug Fix: Delete Daily Button
Delete Daily button brings up the error invalid index, seems to be an issue with how index 0 is parsed. Also because this doesn't use /removedaily you can't actually fully remove a daily from the menu, I changed it so that index 0 does so, and that index 0 is properly read and not passed to 1 every time as I think it's doing so right now?