Commands: Update help FAQ and improve !faq fails
https://www.smogon.com/forums/threads/clean-up-update-error-message-for-faq.3764560/ 
!faq fails no longer send the failed message message in chat, this is very helpful for room auth who are the users of broadcast commands
For whatever reason '!faq all' fails used a replybox? I can't think of a single non dev command that uses visible replyboxes in chat for fails, so changed this for an errorReply.
I cleaned up the wording of the /help faq, and I added some topics that were not there previously.