Hangman: Support optional timer
suggestion: https://www.smogon.com/forums/threads/add-a-timer-to-hangmans.3765293/ 

Tested locally, everything works well. Only super notable thing here is since the hangman constructor needs a number parameter after the hint for timer, I have the integer for "/hangman random" set to 0 so there's no default timer. This has a parse problem though for "/hangman random" because you're supposed to be able to add hints in mass to a single phrase but when you use multiple , the code reads the second string in the hints array as a timer. I separated hints from the hangman class validator to fix that and gave hints its own validator. The validator is light but it works fine for the intended purposes of character count and empty spaces. I also changed args to be called hints to be a bit more clear.