Fix client redirect
For a long time, connecting to non-localhost servers didn't work without
an HTTPS cert. But it's now once again possible. All you need is an
IP with an open port, and a dream.

Syntax is a little different than before. To go to `1.2.3.4` at port `5000`

- before: `http://1.2.3.4-5000.psim.us`
- after: `http://1-2-3-4--5000.insecure.psim.us`

Supporting multiple dots in the URL is too complicated, unfortunately.

The other way still works: `http://1.2.3.4:5000`

And yes, I would 100% recommend that one.