Fixed and Adjusted Abilities
Fixed Mystic Haze - Should be Misty Terrain Now
Adjusted Ejected - Made the ability to be treated like Emergency Exit in terms of when it should be switching out.
Adjusted Sheer Willpower - Made the ability to operate in a similar way as Solar Beam, but without weather.
I will work on the visual stuff tomorrow.