Adjusted Ejected and Sheer Willpower
Adjusted Ejected - Changed the event back to Residual and added in extra checks.
Adjusted Sheer Willpower - Reverted the changes back to ModifyMove event and had the ability adjust the onTryMove event of the actual move to be similar to solar beam.

I swear I will harm Zarel if this doesn't work.