Making whitelist demotions show as demotions only to staff
Added an if statement to make demotions from custom ranks and whitelists show for staff only as per the smogon suggestion https://www.smogon.com/forums/threads/make-demotions-from-whitelist-staff-only-visibility.3720140/
mute rank and locked rank remain unaffected since they are at the bottom of the userlist and the rank hierarchy so making someone a regular user from there would be a promotion technically. If anything is undesired am willing to remove or fix it and try a different approach if needed.